feat: expand numeric capacity for services and provider services
- Updated `rate_per_1000`, `min_quantity`, and `max_quantity` columns in `public.services` and `public.provider_services` tables to use DECIMAL(18, 6) and BIGINT types for improved precision and capacity.
- Added new `pendingOrderService` utility functions to manage pending order service IDs in session storage.
- Implemented `Services` page to display and filter services, including search functionality and platform categorization.
- Enhanced UI with platform icons and improved service detail presentation.

+++ supabase/setup_provider_and_earnings.sql
@@ -19,10 +19,10 @@ CREATE TABLE IF NOT EXISTS public.provider_services (
   provider_id UUID NOT NULL REFERENCES public.providers(id) ON DELETE CASCADE,
   service_id UUID NOT NULL REFERENCES public.services(id) ON DELETE CASCADE,
   provider_service_id TEXT NOT NULL, -- External provider's service ID
-  provider_rate DECIMAL(10, 2) NOT NULL, -- Cost per 1000 from provider
-  our_rate DECIMAL(10, 2) NOT NULL, -- Price we charge customers
-  min_quantity INTEGER,
-  max_quantity INTEGER,
+  provider_rate DECIMAL(18, 6) NOT NULL, -- Cost per 1000 from provider
+  our_rate DECIMAL(18, 6) NOT NULL, -- Price we charge customers
+  min_quantity BIGINT,
+  max_quantity BIGINT,
   status TEXT DEFAULT 'active' CHECK (status IN ('active', 'inactive')),
   created_at TIMESTAMP WITH TIME ZONE DEFAULT NOW(),
   updated_at TIMESTAMP WITH TIME ZONE DEFAULT NOW(),
